requirements.txt
Django>=3.1.0
pytz==2020.5
djangorestframework>=3.13.1